May 19/requirement.txt
selenium==4.12.0
webdriver-manager==4.0.0
beautifulsoup4==4.12.2
google-generativeai==0.4.0
httpx>=0.28.1,<1.0.0
nest-asyncio==1.5.7